Move pubmed logic out
